requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
#connect-box==0.2.8
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
future==0.18.2
idna==2.9
multidict==4.7.6
pycparser==2.20
python-telegram-bot==12.7
PyYAML==5.3.1
six==1.15.0
tornado==6.0.4
yarl==1.4.2
